% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggreg_data}
\alias{aggreg_data}
\title{Simulated aggregated data for Life Years Lost estimation.}
\format{
A data frame with 50 rows and 4 variables:
\describe{
  \item{age}{age}
  \item{new_cases}{number of new cases diagnosed at that specific age}
  \item{deaths}{number of deaths among the diagnosed at that specific age}
  \item{rate}{age-specific mortality rates among the diagnosed}
}
}
\source{
Simulated data
}
\usage{
aggreg_data
}
\description{
A dataset containing age-specific number of new cases, number
of deaths and mortality rates for a simulated disease with onset
after age 40 years. Data is available for ages 40-90 years.
}
\keyword{datasets}
