% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ref}
\alias{pop_ref}
\title{Aggregated data for Life Years Lost estimation.}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{age}{age}
  \item{survival}{survival probability at that specific age}
  \item{mortality_rates}{age-specific mortality rates}
}
}
\source{
Statistics Danmark (https://www.dst.dk/en/Statistik/emner/befolkning-og-valg/)
}
\usage{
pop_ref
}
\description{
A dataset containing age-specific survival probability and mortality
rates for Danish women in years 2017-2018.
}
\keyword{datasets}
