% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu_data}
\alias{simu_data}
\title{Simulated population for Life Years Lost estimation.}
\format{A data frame with 100000 rows and 6 variables:
\describe{
  \item{id}{unique identifier of each person}
  \item{age_start}{age at start of follow-up (0 for all individuals)}
  \item{age_death}{age at end of follow-up (death or censoring)}
  \item{death}{logical variable (\code{TRUE} = death / \code{FALSE} = censoring)}
  \item{cause_death}{factor variable with 3 levels: \code{"Alive"} (for those
  censored) and \code{"Natural"} and \code{"Unnatural"} (for those dying of natural
  and unnatural causes of death, respectively)}
  \item{age_disease}{age at developing a specific disease or condition for those
  32,391 individuals that develop the disease (missing for the remaining 67,609)}
}}
\source{
Simulated data
}
\usage{
simu_data
}
\description{
A dataset containing age and cause of death, as well as
age at disease diagnosis (or start of a condition) for
100,000 simulated persons.
}
\keyword{datasets}
